+++ NEWS.md
@@ -1,3 +1,29 @@
+# CHANGES IN listr VERSION 0.1.1
+
+## NEW FEATURES
+
+- `getByPath()`:
+  Retrieve values by name-based or position based path
+- new fields in structure table:
+  - `path`: names-based path
+  - `ppath`: position-based path
+
+## BUG FIXES
+
+- # 1
+
+## MAJOR CHANGES
+
+- A: `getByPath()`
+- A: alternative way of retrieving structure table in `getRawStructure()`
+- C: field `pindex` --> `oindex` (not *parent*, but *outer*)
+
+## MINOR CHANGES
+
+## MISC
+
+-----
+
 # CHANGES IN listr VERSION 0.1
 
 ## NEW FEATURES